v14 fram implementation

+++ modules/Userapi/Http/Controllers/UserapiControllerV12.php
@@ -85,11 +85,11 @@ class UserapiControllerV12 extends Controller
                 }
                 $referral_by = 0;
                 if($referral_code != "") {
-                    $checkReferralCode = UserReferral::where('referral_code',$referral_code)->first();
+                    $checkReferralCode = User::where('ref_code',$referral_code)->first();
                     if(empty($checkReferralCode)) {
                         return response()->json(['status'=>false,'message'=>'Invalid referral code']);
                     }
-                    $referral_by = $checkReferralCode->user_id;
+                    $referral_by = $checkReferralCode->id;
                 }
                 $checkUser->name = $name;
                 $checkUser->country_code = $country_code;
@@ -5679,6 +5679,9 @@ class UserapiControllerV12 extends Controller
         $socialLogin = SocialLogin::where('user_id', $user_id)->where('type', $request->type)->where('profile_id', $profile_id)->first();
         if(empty($socialLogin)) {
             $socialLogin = new SocialLogin;
+            if($request->login_for == "profile") {
+                $socialLogin->profile_added = 1;
+            }
         }
         else {
             if($socialLogin->profile_added == 0) {
@@ -5724,21 +5727,28 @@ class UserapiControllerV12 extends Controller
 
                     array_push($pageData, $data);
                 }
+
                 $linkedPages = LinkedInPage::where('user_id', $user_id)->where('profile_id', $profile_id)->pluck('page_id')->toArray();
                 $data = ['pageList' => $pageData, "linkedPages" => $linkedPages, 'profile_id' => $profile_id];
                 if(count($pageData) == 0) {
                     return response()->json(['status' => false,'message'=>'No page found', 'data' => $data]);
                 }
+                
                 return response()->json(['status' => true,'message'=>'Page List', 'data' => $data, 'profile_id' => $profile_id]);
             }
             if($request->type == "facebook") {
                 $pages = Helper::getUserFacebookPages($auth_token);
                 $pageData = $pages->data;
+                $pageList = array();
+                foreach($pageData as $page) {
+                    $page->image = "https://graph.facebook.com/" . $page->id . "/picture?type=large";
+                    array_push($pageList, $page);
+                }
                 if(count($pageData) == 0) {
-                    return response()->json(['status' => false,'message'=>'No page found', 'data' => $pageData]);
+                    return response()->json(['status' => false,'message'=>'No page found', 'data' => $pageList]);
                 }
                 $facebookPages = FacebookPage::where('user_id', $user_id)->where('profile_id', $profile_id)->pluck('page_id')->toArray();
-                $data = ['pageList' => $pageData, "facebookPages" => $facebookPages, 'profile_id' => $profile_id];
+                $data = ['pageList' => $pageList, "facebookPages" => $facebookPages, 'profile_id' => $profile_id];
                 return response()->json(['status' => true,'message'=>'Page List', 'data' => $data]);
             }
         }
@@ -5873,7 +5883,8 @@ class UserapiControllerV12 extends Controller
                 $page_id = $page->id;
                 $auth_token = $page->access_token;
                 $page_name = $page->name;
-                $page_photo = $default_profile_photo;
+                $page_photo = "https://graph.facebook.com/" . $page_id . "/picture?type=large";
+                // $page_photo = $default_profile_photo;
                 $checkPage = FacebookPage::where('user_id', $user_id)->where('profile_id', $profile_id)->where('page_id',$page_id)->first();
                 if(empty($checkPage)) {
                     $checkPage = new FacebookPage;
@@ -5899,13 +5910,18 @@ class UserapiControllerV12 extends Controller
         $profile_id = $request->profile_id;
         $checkProfile = SocialLogin::where('type', $profile_type)->where('profile_id', $profile_id)->where('user_id', $user_id)->first();
         if(!empty($checkProfile)) {
-            $checkPages = LinkedInPage::where('user_id', $user_id)->where('profile_id', $profile_id)->first();
-            if(empty($checkPages)) {
-                $checkProfile->delete();
+            if($profile_type == "linkedin") {
+                $checkPages = LinkedInPage::where('user_id', $user_id)->where('profile_id', $profile_id)->first();
+                if(empty($checkPages)) {
+                    $checkProfile->delete();
+                }
+                else {
+                    $checkProfile->profile_added = 0;
+                    $checkProfile->save();
+                }
             }
             else {
-                $checkProfile->profile_added = 0;
-                $checkProfile->save();
+                $checkProfile->delete();
             }
             return response()->json(['status' => true,'message'=>'Profile removed']);
         }
@@ -5919,6 +5935,20 @@ class UserapiControllerV12 extends Controller
         }
         $page_id = $request->page_id;
         $checkPage = LinkedInPage::where('user_id', $user_id)->where('page_id', $page_id)->first();
+        if(!empty($checkPage)) {
+            $checkPage->delete();
+            return response()->json(['status' => true,'message'=>'Page removed']);
+        }
+        return response()->json(['status' => false,'message'=>'Page not found']);
+    }
+
+    public function removeSocialPageFacebook(Request $request) {
+        $user_id = $this->get_userid($request->token);
+        if($user_id == 0){
+            return response()->json(['status'=>false,'message'=>'user not valid']);
+        }
+        $page_id = $request->page_id;
+        $checkPage = FacebookPage::where('user_id', $user_id)->where('page_id', $page_id)->first();
         if(!empty($checkPage)) {
             $checkPage->delete();
             return response()->json(['status' => true,'message'=>'Page removed']);
@@ -6336,7 +6366,7 @@ class UserapiControllerV12 extends Controller
             return response()->json(['status'=>false,'message'=>'post you are looking for is not availabe']);
         }
 
-        $getPost = DB::table('schedule_post')->select('sp_media_path')->where('sp_user_id', $user_id)->where('sp_id','=',$input['id'])->first();
+        $getPost = DB::table('schedule_post')->select('sp_media_path', 'is_posted')->where('sp_user_id', $user_id)->where('sp_id','=',$input['id'])->first();
 
         if($getPost){
             if($getPost->is_posted == 1) {
